Clean up WizardManager and NewSiteCreationMainVM
