+++ WordPress/src/main/java/org/wordpress/android/modules/ViewModelModule.java
@@ -5,9 +5,9 @@ import android.arch.lifecycle.ViewModelProvider;
 
 import org.wordpress.android.ui.JetpackRemoteInstallViewModel;
 import org.wordpress.android.ui.reader.viewmodels.ReaderPostListViewModel;
+import org.wordpress.android.ui.sitecreation.NewSiteCreationMainVM;
 import org.wordpress.android.ui.sitecreation.segments.NewSiteCreationSegmentsViewModel;
 import org.wordpress.android.ui.sitecreation.verticals.NewSiteCreationVerticalsViewModel;
-import org.wordpress.android.util.wizard.SiteCreationMainVM;
 import org.wordpress.android.viewmodel.ViewModelFactory;
 import org.wordpress.android.viewmodel.ViewModelKey;
 import org.wordpress.android.viewmodel.activitylog.ActivityLogDetailViewModel;
@@ -93,8 +93,8 @@ abstract class ViewModelModule {
 
     @Binds
     @IntoMap
-    @ViewModelKey(SiteCreationMainVM.class)
-    abstract ViewModel siteCreationMainVM(SiteCreationMainVM viewModel);
+    @ViewModelKey(NewSiteCreationMainVM.class)
+    abstract ViewModel newSiteCreationMainVM(NewSiteCreationMainVM viewModel);
 
     @Binds
     abstract ViewModelProvider.Factory provideViewModelFactory(ViewModelFactory viewModelFactory);
